Move widgets to oterm.app.widgets
